# apps/api/openapi/sprint-03-attendance-shift-operations.yaml
openapi: 3.1.0
info:
  title: PhoenixHRMS Sprint 03 Attendance and Shift Operations API
  version: 1.0.0
  description: Tenant-scoped attendance capture, calculation, policy, holiday configuration, shift scheduling, assignment, and roster APIs implemented during Sprint 03.
  license:
    name: Proprietary
    identifier: LicenseRef-PhoenixHRMS-Proprietary
servers:
  - url: https://api.phoenixhrms.test/api/v1
security:
  - bearerAuth: []
paths:
...
  /attendance/holiday-calendars/{holidayCalendarId}/holidays:
    post:
      operationId: createHoliday
      summary: Create a holiday inside a tenant holiday calendar
      parameters:
        - $ref: '#/components/parameters/HolidayCalendarId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HolidayRequest'
      responses:
        '201':
          description: Holiday created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Holiday'
        '401':
          $ref: '#/components/responses/UnauthorizedErrorResponse'
        '403':
          $ref: '#/components/responses/ForbiddenErrorResponse'
        '404':
          $ref: '#/components/responses/NotFoundErrorResponse'
        '422':
          $ref: '#/components/responses/ValidationErrorResponse'
  /attendance/holiday-calendars/{holidayCalendarId}/holidays/{holidayId}:
    patch:
      operationId: updateHoliday
      summary: Update a holiday inside a tenant holiday calendar
      parameters:
        - $ref: '#/components/parameters/HolidayCalendarId'
        - $ref: '#/components/parameters/HolidayId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HolidayRequest'
      responses:
        '200':
          description: Holiday updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Holiday'
        '401':
          $ref: '#/components/responses/UnauthorizedErrorResponse'
        '403':
          $ref: '#/components/responses/ForbiddenErrorResponse'
        '404':
          $ref: '#/components/responses/NotFoundErrorResponse'
        '422':
          $ref: '#/components/responses/ValidationErrorResponse'
...
components:
  parameters:
...
    HolidayCalendarId:
      in: path
      name: holidayCalendarId
      required: true
      schema:
        type: integer
    HolidayId:
      in: path
      name: holidayId
      required: true
      schema:
        type: integer
...
  responses:
    UnauthorizedErrorResponse:
      description: Authentication is required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ForbiddenErrorResponse:
      description: Permission denied
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFoundErrorResponse:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ValidationErrorResponse:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          const: true
        message:
          type: string
        data: {}
      required: [success, message, data]
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          const: false
        message:
          type: string
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
      required: [success, message, errors]
...
    Holiday:
      type: object
      properties:
        id:
          type: integer
        holiday_calendar_id:
          type: integer
        name:
          type: string
        holiday_date:
          type: string
          format: date
        type:
          type: string
          enum: [national, regional, company, optional]
        is_optional:
          type: boolean
        description:
          type: [string, 'null']
        created_at:
          type: [string, 'null']
          format: date-time
        updated_at:
          type: [string, 'null']
          format: date-time
      required: [id, holiday_calendar_id, name, holiday_date, type, is_optional]
    HolidayRequest:
      type: object
      properties:
        name:
          type: string
        holiday_date:
          type: string
          format: date
        type:
          type: string
          enum: [national, regional, company, optional]
        is_optional:
          type: boolean
        description:
          type: [string, 'null']
      required: [name, holiday_date, type, is_optional]
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT